% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_3_euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
Greatest common divisor of a and b \code{a} and \code{b}.
}
\description{
In this algorithm, the quotient and remainder are found by repeated subtraction, until the last remainder that is grater than zero. 
Euclidean division reduces all the steps between two exchanges into a single step.
}
\details{
greatest common divisor two numbers by euclidean algorithm.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
