% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_data.r
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{dataframe wiki_graph}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{v1}{wiki_graph}
  \item{v2}{wiki_graph}
  \item{w}{weight of the wiki_graph}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
wiki_graph
}
\description{
dataframe wiki_graph
}
\keyword{datasets}
