% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_3_euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
Greatest common divisor of a and b \code{a} and \code{b}.
}
\description{
(a,b)=(b,r) if a=b*q+r and r is remaining.
}
\details{
greatest common divisor two numbers by euclidean algorithm.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
